implementing loadable structure
