+++ package.json
@@ -5,6 +5,7 @@
   "dependencies": {
     "@emotion/react": "^11.9.3",
     "@emotion/styled": "^11.9.3",
+    "@loadable/component": "^5.15.2",
     "@mui/material": "^5.8.7",
     "@react-three/drei": "^9.17.0",
     "@react-three/fiber": "^8.0.27",
