+++ src/App.css
@@ -1,137 +1,86 @@
-.navbar {
-    margin-top: 20px;
+@import url('https://fonts.googleapis.com/css2?family=Work+Sans:ital,wght@0,100;0,400;0,900;1,100;1,400;1,900&display=swap');
+
+:root {
+  --black: rgba(0, 0, 0);
+  --white-65: rgba(255, 255, 255, 0.65);
+  --white-85: rgba(255, 255, 255, 0.85);
+  --white: rgba(255, 255, 255);
 }
 
-.navbar img,
-.footer img {
-    height: 30px;
+* {
+  box-sizing: border-box;
+  font-family: 'Work Sans', sans-serif;
+  margin: 0;
 }
 
-.container {
-    margin-inline: 90px;
+body {
+  background: var(--black);
+  overflow-x: hidden;
+  -webkit-font-smoothing: antialiased;
+  -moz-osx-font-smoothing: grayscale;
 }
 
-.tool {
-    display: grid;
-    grid-template: 'cube cube input score'
-                    'cube cube input code';
-    gap: 30px;
-    margin-bottom: 240px;
-    margin-top: 60px;
-    min-height: 650px;
+h1,
+h2,
+h3,
+h4,
+h5 {
+  color: #ffffff;
+  font-weight: 900;
+  text-transform: uppercase;
 }
 
-.step {
-    background: #000;
-    border: 1px solid #fff;
-    padding: 30px;
+p,
+span {
+  color: var(--white-85);
 }
 
-.fa,
-.step .svg-inline--fa {
-    height: 18px;
-    margin-block: auto;
+p,
+a,
+span {
+  font-weight: 400;
 }
 
-.cube {
-    grid-area: cube;
-    min-width: 500px;
-    padding: 0px;
+a,
+label {
+  color: #fff;
+  font-weight: 900;
+  text-decoration: none;
 }
 
-.input {
-    grid-area: input;
+a:hover,
+a:focus,
+a:active {
+  color: var(--white-85);
 }
 
-.score {
-    grid-area: score;
-    min-width: 300px;
+pre, 
+pre div {
+  padding-left: 0px !important;
 }
 
 .code {
-    grid-area: code;
-    min-width: 300px;
-    font-variant: tabular-nums;
+  color: var(--white-85);
 }
 
-h1 {
-    margin-top: 80px;
-    margin-bottom: 20px;
-    font-size: 64px;
+/* Scrollbar */
+::-webkit-scrollbar {
+  height: 4px;
+  width: 4px;
 }
 
-h2 {
-    margin-bottom: 20px;
-    font-size: 48px;
+/* Track */
+::-webkit-scrollbar-track {
+  background: var(--black);
 }
 
-h3 {
-    margin-bottom: 20px;
-    font-size: 36px;
+/* Handle */
+::-webkit-scrollbar-thumb {
+  background: var(--white-65);
+  border-radius: 2px;
 }
 
-.lead {
-    font-size: 24px;
-    max-width: 850px;
-}
-
-.grey {
-    display: grid;
-    grid-template-columns: 3fr 6fr;
-    min-height: 750px;
-}
-
-.grey .step {
-    height: fit-content;
-    margin-top: 80px;
-    z-index: 1;
-    position: relative;
-}
-
-.grey .step p:nth-child(3) {
-    margin-top: 20px;
-}
-
-.grey .goo {
-    position: relative;
-}
-
-.grey .goo .step {
-    position: absolute;
-    right: 0;
-    top: 0%;
-    bottom: 50%;
-    left: -30%;
-    height: 100%;
-    margin-top: 0px;
-    z-index: 0;
-}
-
-.footer {
-    display: grid;
-    grid-template-columns: 6fr 6fr;
-    margin-bottom: 20px;
-    margin-top: 120px;
-}
-
-.footer p {
-    text-align: right;
-}
-
-.colors {
-    max-height: 175px;
-    overflow-y: scroll;
-}
-
-input {
-    color: white;
-    text-transform: uppercase;
-}
-
-.color-mode {
-    grid-column: 1 / span 4;
-}
-
-.gradient-slider {
-    grid-column: 1 / span 4;
+/* Handle on hover */
+::-webkit-scrollbar-thumb:hover {
+  background: var(--white);
 }
