+++ src/Home/Home.css
@@ -1,0 +1,137 @@
+.navbar {
+    margin-top: 20px;
+}
+
+.navbar img,
+.footer img {
+    height: 30px;
+}
+
+.container {
+    margin-inline: 90px;
+}
+
+.tool {
+    display: grid;
+    grid-template: 'cube cube input score'
+                    'cube cube input code';
+    gap: 30px;
+    margin-bottom: 240px;
+    margin-top: 60px;
+    min-height: 650px;
+}
+
+.step {
+    background: #000;
+    border: 1px solid #fff;
+    padding: 30px;
+}
+
+.fa,
+.step .svg-inline--fa {
+    height: 18px;
+    margin-block: auto;
+}
+
+.cube {
+    grid-area: cube;
+    min-width: 500px;
+    padding: 0px;
+}
+
+.input {
+    grid-area: input;
+}
+
+.score {
+    grid-area: score;
+    min-width: 300px;
+}
+
+.code {
+    grid-area: code;
+    min-width: 300px;
+    font-variant: tabular-nums;
+}
+
+h1 {
+    margin-top: 80px;
+    margin-bottom: 20px;
+    font-size: 64px;
+}
+
+h2 {
+    margin-bottom: 20px;
+    font-size: 48px;
+}
+
+h3 {
+    margin-bottom: 20px;
+    font-size: 36px;
+}
+
+.lead {
+    font-size: 24px;
+    max-width: 850px;
+}
+
+.grey {
+    display: grid;
+    grid-template-columns: 3fr 6fr;
+    min-height: 750px;
+}
+
+.grey .step {
+    height: fit-content;
+    margin-top: 80px;
+    z-index: 1;
+    position: relative;
+}
+
+.grey .step p:nth-child(3) {
+    margin-top: 20px;
+}
+
+.grey .goo {
+    position: relative;
+}
+
+.grey .goo .step {
+    position: absolute;
+    right: 0;
+    top: 0%;
+    bottom: 50%;
+    left: -30%;
+    height: 100%;
+    margin-top: 0px;
+    z-index: 0;
+}
+
+.footer {
+    display: grid;
+    grid-template-columns: 6fr 6fr;
+    margin-bottom: 20px;
+    margin-top: 120px;
+}
+
+.footer p {
+    text-align: right;
+}
+
+.colors {
+    max-height: 175px;
+    overflow-y: scroll;
+}
+
+input {
+    color: white;
+    text-transform: uppercase;
+}
+
+.color-mode {
+    grid-column: 1 / span 4;
+}
+
+.gradient-slider {
+    grid-column: 1 / span 4;
+}
