+++ src/OpenGraph/OpenGraph.css
@@ -1,0 +1,4 @@
+* {
+    box-sizing: border-box;
+    margin: 0;
+}
