+++ src/OpenGraph/OpenGraph.js
@@ -1,3 +1,5 @@
+import "./OpenGraph.css";
+
 const OpenGraph = () => { 
     return (
         <h1>Test</h1>
