+++ src/index.js
@@ -3,9 +3,8 @@ import ReactDOM from 'react-dom/client';
 
 import App from './App';
 
-import './index.css';
-
 const root = ReactDOM.createRoot(document.getElementById('root'));
+
 root.render(
   <React.StrictMode>
     <App />
